fix whammy

+++ src/shared/main/shared_main.cpp
@@ -431,7 +431,7 @@ int8_t handle_calibration_mouse(int8_t previous, int32_t orig_val, int16_t min, 
     return handle_calibration_xbox(previous << 8, orig_val, min, max, center, deadzone) >> 8;
 }
 
-uint8_t handle_calibration_ps3_360_trigger(uint8_t previous, uint16_t orig_val, int16_t min, int16_t multiplier, uint16_t deadzone) {
+uint8_t handle_calibration_ps3_360_trigger(uint8_t previous, uint16_t orig_val, uint32_t min, int16_t multiplier, uint16_t deadzone) {
     return handle_calibration_xbox_one_trigger(previous << 2, orig_val, min, multiplier, deadzone) >> 2;
 }
